feat(test): add dedicated tests for WorkoutSessionStorage fallback
Adds comprehensive unit and integration tests for the `WorkoutSessionStorage` class, specifically targeting the IndexedDB to localStorage fallback logic.

This change introduces two distinct test suites:
- `workout-session-storage.test.ts`: An integration test suite using `fake-indexeddb` to verify the primary path where IndexedDB is available.
- `workout-session-storage.fallback.test.ts`: A unit test suite using manual mocks to verify the fallback path where IndexedDB fails.

The tests cover all core methods of the `WorkoutSessionStorage` class for both the primary and fallback storage mechanisms, ensuring data is correctly persisted and retrieved in both scenarios. This addresses the lack of test coverage for this critical logic and resolves previous test failures.
